extract java beans spec to separate file
